feat: Use floating point fee rate (#3066)
# Description

---------

+++ crates/bitcoin-da/src/monitoring.rs
@@ -50,7 +50,7 @@ pub enum TxStatus {
     #[serde(rename_all = "camelCase")]
     InMempool {
         /// Base fee rate.
-        base_fee: u64,
+        base_fee: f64,
         /// Timestamp.
         timestamp: u64,
         /// Block height when transaction entered pool
@@ -727,7 +727,7 @@ impl MonitoringService {
         } else {
             match self.client.get_mempool_entry(&tx_result.info.txid).await {
                 Ok(entry) => {
-                    let base_fee = entry.fees.base.to_sat();
+                    let base_fee = entry.fees.base.to_sat() as f64;
                     TxStatus::InMempool {
                         base_fee,
                         timestamp: get_timestamp(),
